Migrate squads module to new structure
