+++ modules/region_assignment/command-assign-na.js
@@ -3,9 +3,9 @@
 const
     config = require('config'),
 
-    CommandAssignment = require('./command-assignment');
+    CommandAssignmentBase = require('./command-assignment-base');
 
-class CommandAssignNA extends CommandAssignment {
+class CommandAssignNA extends CommandAssignmentBase {
     constructor(module) {
         super(module);
