+++ modules/region_assignment/command-remove-eu.js
@@ -3,9 +3,9 @@
 const
     config = require('config'),
 
-    CommandAssignment = require('./command-assignment');
+    CommandAssignmentBase = require('./command-assignment-base');
 
-class CommandRemoveEU extends CommandAssignment {
+class CommandRemoveEU extends CommandAssignmentBase {
     constructor(module) {
         super(module);
